Fix routing, reorganize project structure
