+++ frontend/src/pages/calculatedResults/index.js
@@ -1,0 +1,15 @@
+import React from 'react'
+import "./style.css";
+
+const CalculatedResults = () => {
+
+  return (
+      <div style={{display: 'flex', justifyContent:'center', alignItems: 'center', height: '90vh'}}>
+        <h1>
+          Placeholder for calculated results
+        </h1>
+    </div>
+  );
+};
+
+export default CalculatedResults
