+++ frontend/src/pages/calculator/index.js
@@ -1,13 +1,14 @@
 import React from 'react'
+import "./style.css";
 
 const Calculator = () => {
   return (
     <div style={{display: 'flex', justifyContent:'center', alignItems: 'center', height: '90vh'}}>
-        <h1>
-            Placeholder for Calculator page
-        </h1>
-    </div>
-  )
-}
+      <h1>
+        Placeholder for calculator
+      </h1>
+  </div>
+);
+};
 
 export default Calculator
