--- frontend/src/pages/index.js
@@ -1,14 +1,0 @@
-import React from 'react'
-import { NavBtn, NavBtnLink } from '../components/navbar/NavbarElements';
-
-const Home = () => {
-  return (
-    <div style={{display: 'flex', justifyContent:'center', alignItems: 'center', height: '90vh'}}>
-        <h1>
-            Placeholder for Homepage
-        </h1>
-    </div>
-  );
-};
-
-export default Home;
