--- frontend/src/pages/usefulTerms.js
@@ -1,13 +1,0 @@
-import React from 'react'
-
-const usefulTerms = () => {
-  return (
-    <div style={{display: 'flex', justifyContent:'center', alignItems: 'center', height: '90vh'}}>
-        <h1>
-            Placeholder for Useful Terms page
-        </h1>
-    </div>
-  );
-};
-
-export default usefulTerms
