+++ frontend/src/pages/usefulTerms/usefulTerms.js
@@ -1,13 +1,13 @@
 import React from 'react'
 
-const HowToCalculateInstallments = () => {
+const UsefulTerms = () => {
   return (
     <div style={{display: 'flex', justifyContent:'center', alignItems: 'center', height: '90vh'}}>
         <h1>
-        Placeholder for How we calculate installments page
+            Placeholder for useful terms
         </h1>
     </div>
   );
 };
 
-export default HowToCalculateInstallments
+export default UsefulTerms
