+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "victory": "^36.3.1"
+  }
+}
